+++ src/QNEthernetFrame.cpp
@@ -12,6 +12,8 @@
 // C++ includes
 #include <algorithm>
 
+#include <util/atomic.h>
+
 #include "QNEthernet.h"
 #include "lwip/prot/ieee.h"
 
@@ -65,10 +67,9 @@ err_t EthernetFrameClass::recvFunc(struct pbuf *p, struct netif *netif) {
   return ERR_OK;
 }
 
-EthernetFrameClass::EthernetFrameClass() {
-  for (Frame &f : inBuf_) {
-    f.data.reserve(maxFrameLen());
-  }
+EthernetFrameClass::EthernetFrameClass()
+    : inBuf_(1) {
+  setReceiveQueueSize(1);
 }
 
 // --------------------------------------------------------------------------
@@ -216,6 +217,38 @@ int EthernetFrameClass::availableForWrite() {
   // First cast to something we know is the same size as size_t
   return std::max(
       static_cast<ssize_t>((maxFrameLen() - 4) - outFrame_.data.size()), 0);
+}
+
+void EthernetFrameClass::setReceiveQueueSize(size_t size) {
+  if (size < 1) {
+    size = 1;
+  }
+
+  size_t oldSize = inBuf_.size();
+
+  // Keep all the newest elements
+  ATOMIC_BLOCK(ATOMIC_RESTORESTATE) {
+    if (size <= inBufSize_) {
+      // Keep all the newest frames
+      if (inBufTail_ != 0) {
+        size_t n = (inBufTail_ + (inBufSize_ - size)) % inBuf_.size();
+        std::rotate(inBuf_.begin(), inBuf_.begin() + n, inBuf_.end());
+      }
+      inBuf_.resize(size);
+      inBufHead_ = 0;
+      inBufSize_ = size;
+    } else {
+      if (inBufTail_ != 0) {
+        std::rotate(inBuf_.begin(), inBuf_.begin() + inBufTail_, inBuf_.end());
+      }
+      inBuf_.resize(size);
+      inBufHead_ = inBufSize_;
+      for (size_t i = oldSize; i < size; i++) {
+        inBuf_[i].data.reserve(maxFrameLen());
+      }
+    }
+    inBufTail_ = 0;
+  }
 }
 
 }  // namespace network